change settings

+++ aga_transformers/models/t5/modeling_t5_slides.py
@@ -1040,11 +1040,11 @@ class FlaxT5EfficientBlockGraphSelfAttention(nn.Module):
         #TODO: compute these values (or retrieve them from the input graph)
         
         if self.has_variable("graph", "receivers"):
-            slide_start_for_blocks = self.variables["graph"]["slide_start_for_blocks"] #array of slide start indices
+            slide_start_for_blocks = self.variables["graph"]["slide_start_for_blocks"].astype(jnp.int32) #array of slide start indices
             n_slides_total = self.variables["graph"]["n_slides"] #int = number of slides in total
         else:
             slide_start_for_blocks = jnp.array([[0 for _ in range(math.ceil(8192 / block_len))]*batch_size], dtype=jnp.uint16)
-            n_slides_total = jnp.zeros((batch_size,), dtype=jnp.uint16)
+            n_slides_total = jnp.zeros((batch_size,), dtype=jnp.int32)
         
         n_slides_context = 8 #static int = number of slides in the context window
         n_global_tokens = max_slides + n_document_tokens # was 12, static value that should be >= n_document_tokens + n_slides.max()
